requirements.txt
python-telegram-bot[job-queue]
gspread
google-auth
Pillow
python-dotenv
groq
anthropic
